revert read and write lp files to qp
